% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds-methods.R
\name{buildCDS}
\alias{buildCDS}
\title{Build coding segments}
\usage{
buildCDS(object, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{verbose}{Whether to print out messages (Default: FALSE)}
}
\value{
Updated factRObject.
The GTF GenomicRanges object will now contain "CDS" type features.

The `cds` variable from the transcript metadata will return "yes" for coding
transcripts and "no" for non-coding transcripts
}
\description{
Constructs CDS information on transcripts from custom annotation
using reference-based approach.
}
\details{
This function will firstly identify custom transcripts with identical
exon structures as those in the reference annotation. If these reference
transripts contain CDS segments, the coordinates will be passed to its paired-
custom transcript. The putative start codon of the remaining transcripts
will be the first ATG sequence that is in-frame with the coding sequence
of reference transcripts and an in-frame stop codon will be determined.
}
\examples{
## Load sample data
data(factRsample)

## Build coding sequences
factRsample <- buildCDS(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
