% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{createfactRObject}
\alias{createfactRObject}
\title{Construct a factRObject}
\usage{
createfactRObject(
  gtf,
  reference,
  use_own_annotation = NULL,
  use_own_genome = NULL,
  counts,
  psi,
  assays = "transcript",
  verbose = FALSE
)
}
\arguments{
\item{gtf}{path to custom GTF file.}

\item{reference}{a character value of supported reference ID. Run
\code{\link{ListSupportedGenomes}} for a list of supported reference. Input
can also be a reference species name ("Mouse" or "Mus musculus or "Mmusculus").
To use your own annotation and genome, leave this parameter blank and provide
inputs to `use_own_annotation` and `use_own_genome`}

\item{use_own_annotation}{can be a (1) GenomicRanges object containing
reference transcript annotation, (2) path to local GTF file, (3) AnnotationHub
ID or (4) URL to an online GTF file.}

\item{use_own_genome}{can be a (1) BSgenome or DNAStringSet object containing
genome sequence, (2) path to local genome fasta file, (3) AnnotationHub
ID or (4) URL to an online genome fasta file.}

\item{verbose}{if TRUE, show progress messages}
}
\value{
factR object
}
\description{
Construct a factRObject
}
\examples{
# Create factRObject using sample GTF and supported reference
gtf <- system.file("extdata", "sc_merged_sample.gtf.gz", package = "factR")
obj <- createfactRObject(gtf, "vM25")

# Create factRObject using custom reference
library(BSgenome.Mmusculus.UCSC.mm10)
obj <- createfactRObject(gtf, use_own_annotation = "AH49547", use_own_genome = Mmusculus)
}
