% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factRObject-class.R
\name{exportGTF,factR-method}
\alias{exportGTF,factR-method}
\alias{exportTable,factR-method}
\alias{exportAll,factR-method}
\title{Exporting data from factRObject}
\usage{
\S4method{exportGTF}{factR}(object, out = getwd())

\S4method{exportTable}{factR}(object, out = getwd(), data = activeSet(object))

\S4method{exportAll}{factR}(object, path = getwd())
}
\arguments{
\item{object}{factRObject}

\item{out}{Path to output directory or path to output file. If the former,
the output file will be named "factR.gtf"}

\item{data}{Set metadata to export Can be "gene", "transcript" or "AS".}

\item{path}{Path to output directory or path to output file.}
}
\description{
Exporting data from factRObject
}
