% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnmd-methods.R
\name{factR-NMD}
\alias{factR-NMD}
\alias{testASNMDevents,factR-method}
\title{NMD predictions at transcript and exon levels}
\usage{
\S4method{predictNMD}{factR}(object, NMD_threshold = 50, verbose = FALSE)

\S4method{testASNMDevents}{factR}(object, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{NMD_threshold}{Minimum distance between PTC and downstream exon-exon junction to trigger NMD (Default: 50)}

\item{verbose}{Whether to print out messages (Default: FALSE)}
}
\value{
factRObject with updated metadata

For `predictNMD`, 4 additional variables are added to the transcript metadata:
\itemize{
 \item{nmd: }{"yes" or "no" value as to whether the transcript is NMD-sensitive}
 \item{stop_to_lastEJ: }{Integer of the distance between the first stop codon
 to the last exon-junction. Positive values indicate that the stop codon
 is upstream of the EJ while negative values indicate that the stop codon
 is downstrea of the EJ }
 \item{num_of_downEJs: }{Number of EJs downstream of the first stop codon}
 \item{3'UTR_length: }{Length of the 3'UTR}
}

For `testASNMDevents`, 2 additional variables are added to the AS metadata:
\itemize{
 \item{ASNMDtype: }{Type of AS-NMD event. Can be "Repressing" if skipping
 of the exon leads to NMD or "Stimulating" if splicing of the exon leads to
 NMD}
 \item{ASNMD.in.cds: }{Whether or not the event is found within the CDS or UTR}
}
}
\description{
Nonsense-mediated decay is an RNA surveillance mechanism which clears
transcripts harboring a premature stop codon. These transcripts can be
products of alternative splicing events (AS-NMD) which ultimately serve as
a post-transcriptional gene regulation mechanism.

The functions below annotates all transcripts in the factRObject for its
sensitivity to NMD, based on the canonical exon-junction model. Protein-coding
transcripts containing a premature stop codon >50bp upstream of the
last exon junction will be annotated as NMD-sensitive.

Sequentially, the `testASNMDevents` function will pinpoint the alternative
splicing events that lead to NMD. AS-NMD events are categorised as "stimulating"
or "repressing". "Stimulating" events trigger NMD upon its splicing
while "repressing" events trigger NMD upon exon skipping.
}
\examples{
## Load sample factR object and predict CDS segments
data(factRsample)
factRsample <- buildCDS(factRsample)

## Predict transcript-level NMD sensitivities
factRsample <- predictNMD(factRsample)

## Annotate NMD-causing splicing events
factRsample <- testASNMDevents(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
