% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{FindAltSplicing}
\alias{FindAltSplicing}
\title{Find alternative splicing events}
\usage{
FindAltSplicing(object)
}
\arguments{
\item{object}{factRObject}
}
\value{
Updated factRObject
}
\description{
This function identifies all alternative splicing events that
occur in the custom transcriptome.
}
\examples{
data(factRsample)
factRsample <- FindAltSplicing(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
