% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-methods.R
\name{getAAsequence,factR-method}
\alias{getAAsequence,factR-method}
\title{Translate coding sequence}
\usage{
\S4method{getAAsequence}{factR}(object, verbose = FALSE)
}
\arguments{
\item{factR}{}
}
\description{
Translate coding sequence
}
