% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factRObject-class.R
\name{getAAsequence}
\alias{getAAsequence}
\title{Get protein coding sequences}
\usage{
getAAsequence(object, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{verbose}{whether to display workflow progress}
}
\value{
Updated factRObject
}
\description{
This function will translate coding transcript sequences into
amino acid sequences and stores them as a dataframe in the object's
domain slot
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- getAAsequence(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
