% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation-methods.R
\name{getAScons}
\alias{getAScons}
\title{Quantify feature conservation}
\usage{
getAScons(object, db = "phastCons", type = "exon", padding = 50)
}
\arguments{
\item{object}{factR object}

\item{db}{Database to extract sequence conservation. Can be "phastCons" (Default) or "phylop"}

\item{type}{Feature to quantify conservation. Can be one of the following:
\itemize{
 \item{"exon"}{ : Sequence conservation of the entire exon}
 \item{"flanks"}{ : Conservation of sequences flanking exons}
 \item{"upstream"}{ : Conservation of sequences upstream of exons}
 \item{"downstream"}{ : Conservation of sequences downstream of exons}
}}

\item{padding}{Additional width to pad the sequence by. 
For cons_type "exons" and "flanks", padding will be added on both sides.}
}
\value{
factRObject with additional columns in ASE metadata.
}
\description{
Quantify sequence conservation scores of alternatively-spliced exons.
}
\details{
By default, this function quantifies the sequence conservation of the entire
exon together with 50 base-pair flanking sequence using the phastCons database.
Alternatively, this function can also quantify the mean conservation scores
of sequences flanking exons, or sequences upstream/downstream of alternative exons.

The `padding` argument has different meaning depending on the `type` input.
If `type` is "exon", then `padding` refers to the amount of intronic sequences
on both sides of the exon to include in the quantification. If `type` is
"flanks", "upstream" or "downstream", then `padding` refers to the amount
of intronic sequnce to
}
\examples{
data(factRsample)
factRsample <- getAScons(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
