% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gtf.R
\name{ListGenomes}
\alias{ListGenomes}
\title{List factR supported genomes}
\usage{
ListGenomes()
}
\value{
data.frame with a list of factR-supported genomes
}
\description{
List factR supported genomes
}
\examples{
ListGenomes()
}
