% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ListSupportedGenomes}
\alias{ListSupportedGenomes}
\title{List factR supported genomes}
\usage{
ListSupportedGenomes()
}
\value{
data.frame with a list of factR-supported genomes
}
\description{
List factR supported genomes
}
\examples{
ListGenomes()
}
