% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnmd-methods.R
\name{predictNMD,factR-method}
\alias{predictNMD,factR-method}
\alias{testASNMDevents}
\title{Predict NMD-sensitive transcripts}
\usage{
\S4method{predictNMD}{factR}(object, NMD_threshold = 50, verbose = FALSE)

testASNMDevents(object, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{NMD_threshold}{Minimum distance between PTC and downstream exon-exon junction to trigger NMD (Default: 50)}

\item{verbose}{Whether to print out messages (Default: FALSE)}
}
\value{
factRObject with update transcripts metadata

factRObject with updated ASE metadata
}
\description{
Annotates NMD-sensitive transcripts based on Exon-Junction model.
Protein-coding transcrupts with a premature stop codon >50bp upstream of the 
last exon junction will be annotated as NMD-sensitive.

Pinpoints alternative splicing events underlying NMD sensitivity.
Metadata of AS events will be updated to include columns "ASNMDtype" and "ASNMD.in.cds".
Values of "ASNMDtype" can be Repressing (skipping of exon causes transcripts to be NMD-sensitive),
Stimulating (splicing of exon causes transcripts to be NMD-sensitive) or
NA (does not cause transcripts to NMD-sensitive).
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- predictNMD(factRsample)
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- predictNMD(factRsample)
factRsample <- testASNMDevents(factRsample)
}
\seealso{
\code{\link{runfactR}} \code{\link{testASNMDevents}}

\code{\link{runfactR}} \code{\link{predictNMD}}
}
