% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-funcs.R
\name{plotDomains,factR-method}
\alias{plotDomains,factR-method}
\title{Plot domain architecture}
\usage{
\S4method{plotDomains}{factR}(object, ..., ncol = 1)
}
\arguments{
\item{object}{factR object class}

\item{...}{One or more features to plot. Can be the following:
\itemize{
 \item{gene_id: }{ID of gene to plot}
 \item{gene_name: }{Name of gene to plot}
 \item{transcript_id: }{ID of transcript to plot}
}}

\item{ncol}{Dimension of patchworked plot. Applicable if 2 or more features are plotted}
}
\value{
Interactive plotly plot
}
\description{
Plot domain architecture
}
\examples{
data(factRsample)
plotTranscripts(factRsample, "Dab2")
plotTranscripts(factRsample, "Dab2", collapse=TRUE)
plotTranscripts(factRsample, "Dab2", rescale_introns=TRUE)
}
