% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-funcs.R
\name{plotTranscripts,factR-method}
\alias{plotTranscripts,factR-method}
\title{Plot transcript architecture}
\usage{
\S4method{plotTranscripts}{factR}(
  object,
  ...,
  collapse = FALSE,
  rescale_introns = FALSE,
  ncol = 1
)
}
\arguments{
\item{object}{factR object class}

\item{...}{One or more features to plot. Can be the following:
\itemize{
 \item{gene_id: }{ID of gene to plot}
 \item{gene_name: }{Name of gene to plot}
 \item{transcript_id: }{ID of transcript to plot}
 \item{AS_id: }{ID of alternative splicing event (ASxxxxx)}
 \item{coordinate: }{In the form of chromosome:start-end}
}}

\item{collapse}{Whether to collapse all transcripts into one (Default: FALSE)}

\item{rescale_introns}{Whether to rescale intron sizes. Useful for genes with
extremely long introns, this will place emphasis on the exon architectures}

\item{ncol}{Dimension of patchworked plot. Applicable if 2 or more features are plotted}
}
\value{
Interactive plotly plot
}
\description{
Plot transcript architecture
}
\examples{
data(factRsample)
plotTranscripts(factRsample, "Dab2")
plotTranscripts(factRsample, "Dab2", collapse=TRUE)
plotTranscripts(factRsample, "Dab2", rescale_introns=TRUE)
}
