% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factRObject-class.R
\name{predictDomain}
\alias{predictDomain}
\title{Predict functional domains}
\usage{
predictDomain(object, ..., database = "superfamily", ncores = 4)
}
\arguments{
\item{object}{factRObject}

\item{...}{a list of features to perform prediction on. Input
can be a mixture of names from gene_id, gene_name or transcript_id metadata.
If missing, function will predict domains on all proteins in the GTF}

\item{database}{character string specifying the database to search. Currently
support "superfamily" (Default) and "pfam"}

\item{ncores}{number of cores to run function (Default: 4)}
}
\value{
Updated factRObject
}
\description{
This function predicts the functional domains found on
proteins
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- predictNMD(factRsample)
}
