% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-methods.R
\name{predictDomains}
\alias{predictDomains}
\title{Predict protein domains}
\usage{
predictDomains(object, ..., database = "superfamily", ncores = 4)
}
\arguments{
\item{object}{factRObject}

\item{...}{One or more features to display. Can be the following:
\itemize{
 \item{gene_id: }{ID of gene to plot}
 \item{gene_name: }{Name of gene to plot}
 \item{transcript_id: }{ID of transcript to plot}
}}

\item{database}{HMM database to query.}

\item{ncores}{Number of cores to run prediction on}
}
\value{
Updated factRObject
}
\description{
Predict protein domains
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- getAAsequence(factRsample)
factRsample <- predictDomains(factRsample, "Osmr")
}
