% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{PredictNMD}
\alias{PredictNMD}
\title{Predict NMD sensitivity on coding transcripts}
\usage{
PredictNMD(object, NMD_threshold = 50, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{NMD_threshold}{Minimum distance of stop_codon to last exon junction (EJ)
which triggers NMD. Default = 50bp}

\item{verbose}{whether to display workflow progress}
}
\value{
Updated factRObject
}
\description{
This function predicts the sensitivity of protein-coding
transcripts to nonsense-mediated decay.
}
\examples{
data(factRsample)
factRsample <- BuildCDS(factRsample)
factRsample <- PredictNMD(factRsample)
}
\seealso{
\code{\link[factR]{predictNMD}}, \code{\link{runfactR}}
}
