% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{view}
\alias{view}
\alias{txRanges}
\alias{txData}
\title{Preview factR object}
\usage{
view(object, ...)

txRanges(object, ...)

txData(object)
}
\arguments{
\item{object}{factRObject}

\item{...}{Optional: a list of features to view. Input
can be a mixture of names from gene_id, gene_name or transcript_id metadata.
If missing, function will return all features}

\item{in_console}{whether to print custom transcriptome in console}
}
\description{
A set of functions to view the contents of a factR object.
\itemize{
 \item{view(): }{view contents of custom transcriptome in spreadsheet-style format}
 \item{txRanges(): }{prints out transcript ranges}
 \item{txData(): }{prints out transcript metadata}
}
}
\examples{

# Ranges of custom transcriptome can be printed out in several ways:
factRsample
head(factRsample)
head(factRsample, 10)
tail(factRsample)


}
\author{
Fursham Hamid
}
