% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-methods.R
\name{testGeneCorr}
\alias{testGeneCorr}
\title{Assess regulatory potential of AS events}
\usage{
testGeneCorr(object, vst = TRUE, min_n = 3, ...)
}
\arguments{
\item{object}{factR object}

\item{vst}{whether to apply variance stabilization on splicing and expression levels}

\item{min_n}{minimum number of non-NA samples required for correlation testing}

\item{...}{additional arguments parsed to cor.test function}
}
\value{
factRObject with updated ASE metadata
}
\description{
Correlates inclusion levels of exons with gene expression levels
}
\examples{
data(factRsample)
factRsample <- runfactR(factRsample)
}
\seealso{
\code{\link{cor.test}}
}
