% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{BuildCDS}
\alias{BuildCDS}
\title{Build coding sequence on custom transcriptome}
\usage{
BuildCDS(object, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{verbose}{whether to display workflow progress}
}
\value{
Updated factRObject
}
\description{
This function constructs coding sequences on custom transcriptomes.
This is done based on the core functionality of factR's buildCDS function.
}
\examples{
data(factRsample)
factRsample <- BuildCDS(factRsample)
}
\seealso{
\code{\link[factR]{buildCDS}}, \code{\link{RunfactR}}
}
