% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds-methods.R
\name{buildCDS}
\alias{buildCDS}
\title{Build coding segments}
\usage{
buildCDS(object, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{verbose}{Whether to print out messages (Default: FALSE)}
}
\value{
Update factRObject with additional data from buildCDS
}
\description{
Constructs CDS information on transcripts from custom annotation
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
