% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-funcs.R
\name{factR-plottings}
\alias{factR-plottings}
\alias{plotDomains,factR-method}
\title{Plot exon and domain architectures}
\usage{
\S4method{plotTranscripts}{factR}(
  object,
  ...,
  collapse = FALSE,
  rescale_introns = FALSE,
  ncol = 1
)

\S4method{plotDomains}{factR}(object, ..., ncol = 1)
}
\arguments{
\item{object}{factR object class}

\item{...}{One or more features to plot. Can be the following:
\itemize{
 \item{gene_id: }{ID of gene to plot}
 \item{gene_name: }{Name of gene to plot}
 \item{transcript_id: }{ID of transcript to plot}
}}

\item{collapse}{Whether to collapse all transcripts into one (Default: FALSE)}

\item{rescale_introns}{Whether to rescale intron sizes. Useful for genes with
extremely long introns, this will place emphasis on the exon architectures}

\item{ncol}{Dimension of patchworked plot. Applicable if 2 or more features are plotted}
}
\value{
Interactive plotly plot
}
\description{
Plot exon and domain architectures
}
\examples{
## Load sample factRObject
data(factRsample)

## Plot transcript archtectures:
plotTranscripts(factRsample, "Dab2")
plotTranscripts(factRsample, "Dab2", collapse=TRUE)
plotTranscripts(factRsample, "Dab2", rescale_introns=TRUE)

## To plot protein-coding domains, the CDS segments need to be built first:
factRsample <- buildCDS(factRsample)
factRsample <- getAAsequence(factRsample)

## Plotting of domains:
plotDomains(factRsample, "Dab2")
}
