% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factRobject.R
\docType{class}
\name{factRObject-class}
\alias{factRObject-class}
\alias{CreatefactRObject}
\title{Create factRObject}
\usage{
CreatefactRObject(gtf, annotation, genome)
}
\arguments{
\item{gtf}{path to custom GTF file.}

\item{annotation}{a reference transcriptome annotation. Input can be (1) a
GenomicRanges object containing transcriptome GTF information, (2) a path
to a reference GTF file on local computer or (3) a character value representing
one of the genome IDs that we support. Run \code{\link{ListGenomes}} for our
list of supported genomes}

\item{genome}{genome sequence. Input can be (1) a
BSgenome object containing genome sequence, (2) a path
to a reference GTF file on local computer or (3) a character value representing
one of the genome IDs that we support. Run \code{\link{ListGenomes}} for our
list of supported genomes}
}
\value{
A factRObject

factR object
}
\description{
factRObject stores all imported inputs, intermediate objects and results of
factR workflow. To construct a factRObject, you need a custom transcriptome
in GTF format, a reference annotation and a genome sequence. See examples
[NOT COMPLETE]
}
\section{Slots}{

\describe{
\item{\code{ranges}}{a list of GenomicRanges GTF objects}

\item{\code{genome}}{genome sequence}

\item{\code{domains}}{a dataFrame of predicted protein domains}

\item{\code{nmd}}{a list of NMD-related objects}

\item{\code{misc}}{a list of miscellaneous information}

\item{\code{version}}{version of factR this object was built under}
}}

\examples{
library(BSgenome.Mmusculus.UCSC.mm10)
gtf <- system.file("extdata", "sc_merged.gtf.gz", package = "factR")

sample.factr <- CreatefactRObject(gtf, "vM25", Mmusculus)
}
