% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/object-constructor.R
\docType{class}
\name{factRObject-class}
\alias{factRObject-class}
\alias{CreatefactRObject}
\title{Create factRObject}
\usage{
CreatefactRObject(
  gtf,
  reference,
  use_own_annotation = NULL,
  use_own_genome = NULL,
  verbose = FALSE
)
}
\arguments{
\item{gtf}{path to custom GTF file.}

\item{reference}{a character value of supported reference ID. Run
\code{\link{ListSupportedGenomes}} for a list of supported reference. Input
can also be a reference species name ("Mouse" or "Mus musculus or "Mmusculus").
To use your own annotation and genome, leave this parameter blank and provide
inputs to `use_own_annotation` and `use_own_genome`}

\item{use_own_annotation}{can be a (1) GenomicRanges object containing
reference transcript annotation, (2) path to local GTF file, (3) AnnotationHub
ID or (4) URL to an online GTF file.}

\item{use_own_genome}{can be a (1) BSgenome or DNAStringSet object containing
genome sequence, (2) path to local genome fasta file, (3) AnnotationHub
ID or (4) URL to an online genome fasta file.}

\item{verbose}{if TRUE, show progress messages}
}
\value{
A factRObject


}
\description{
factRObject stores all imported inputs, intermediate objects and results of
factR workflow. To construct a factRObject, you need a custom transcriptome
in GTF format, a reference annotation and a genome sequence. See examples
[NOT COMPLETE]
}
\section{Slots}{

\describe{
\item{\code{ranges}}{a list of GenomicRanges GTF objects}

\item{\code{genome}}{genome sequence}

\item{\code{domains}}{a dataFrame of predicted protein domains}

\item{\code{nmd}}{a list of NMD-related objects}

\item{\code{misc}}{a list of miscellaneous information}

\item{\code{version}}{version of factR this object was built under}
}}

\examples{
# Create factRObject using sample GTF and supported reference
gtf <- system.file("extdata", "sc_merged_sample.gtf.gz", package = "factR")
obj <- CreatefactRObject(gtf, "vM25")

# Create factRObject using custom reference
library(BSgenome.Mmusculus.UCSC.mm10)
obj <- CreatefactRObject(gtf, use_own_annotation = "AH49547", use_own_genome = Mmusculus)
}
