% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factRObject-class.R
\docType{class}
\name{factRObject-class}
\alias{factRObject-class}
\title{factRObject}
\description{
A container for storing and processing custom transcriptome.
}
\details{
The factRObject is a representation of custom-assembled transcriptome data.
Coordinate and metadata information from an input transcriptome (GTF) is
parsed and grouped into 3 data Sets which corresponds to the levels of gene (genes),
transcript (txs) and alternative splicing (ase). With the aid of a compatible
reference genome and expression data, functional and
comparative analyses can be performed to better understand the complexity
of a transcriptome.
}
\section{Constructor}{

\describe{
     \item{factRObject can be easily constructed as such:}{
         \code{\code{\link{createfactRObject}}(gtf, reference)}
   }
}
}

\section{Accessors}{

Interact with a factRObject (x) the following ways:
\describe{
   \item{}{
      \code{summary(x): Prints preview of factRObject}
   }
   \item{}{
      \code{head(x): Prints the first 6 metadata entries of the active Set}
   }
   \item{}{
      \code{tail(x): Prints the last 6 metadata entries of the active Set}
   }
   \item{}{
      \code{dim(x): Prints the dimension of the active set (features x samples)}
   }
   \item{}{
      \code{nrow(x): Returns the number of features of the active set}
   }
   \item{}{
      \code{ncol(x): Returns the number of samples}
   }
}

Working with Sets object:
\describe{
   \item{}{
      \code{\code{\link{listSets}}(x): Lists Sets in object}
   }
   \item{}{
      \code{\code{\link{activeSet}}(x): Returns the active Set of object}
   }
   \item{}{
      \code{\code{\link{activeSet}}(x)<-: Change the active Set}
   }
}

Preview Sets metadata as such:
\describe{
   \item{}{
      \code{\code{\link{features}}(x): Displays metadata of active Set}
   }
   \item{}{
      \code{\code{\link{genes}}(x) or \code{\link{gns}}(x): Displays gene metadata}
   }
   \item{}{
      \code{\code{\link{transcripts}}(x) or \code{\link{txs}}(x): Displays transcript metadata}
   }
   \item{}{
      \code{\code{\link{ase}}(x): Displays alternative splicing events metadata}
   }
}

Plot transcripts or domains:
\describe{
   \item{}{
      \code{\code{\link{plotTranscripts}}(x, "gene of interest"): Plots transcripts}
   }
   \item{}{
      \code{\code{\link{plotDomains}}(x, "gene of interest")): Plots domains}
   }
}
}

