% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factRObject-class.R
\docType{class}
\name{factRObject-class}
\alias{factRObject-class}
\title{factRObject}
\description{
A container for storing and processing custom transcriptome.
}
\details{
The factRObject is a representation of custom-assembled transcriptome data.
Coordinate and metadata information from an input transcriptome (GTF) is
parsed and grouped into 3 data sets which corresponds to the levels of gene,
transcript and alternative splicing (AS). With the aid of a compatible
reference genome and expression data, functional and
comparative analyses can be performed to better understand the complexity
of a transcriptome.
}
\section{Constructor}{

\describe{
   \item{factRObject can be easily constructed as such:}{
      \code{createfactRObject(gtf, reference)}
   }
}
}

\section{Accessors}{

Test
}

