% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factRObject-class.R
\name{findAltSplicing}
\alias{findAltSplicing}
\title{Find alternative splicing events}
\usage{
findAltSplicing(object)
}
\arguments{
\item{object}{factRObject}
}
\value{
Updated factRObject
}
\description{
This function identifies all alternative splicing events that
occur in the custom transcriptome.
}
\examples{
data(factRsample)
factRsample <- findAltSplicing(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
