% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-methods.R
\name{getAAsequence,factR-method}
\alias{getAAsequence,factR-method}
\title{Translate coding sequence}
\usage{
\S4method{getAAsequence}{factR}(object, verbose = FALSE)
}
\arguments{
\item{factR}{}
}
<<<<<<< HEAD
\description{
Translate coding sequence
}
=======
\value{
Updated factR object
}
\description{
Translate coding sequence
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- getAAsequence(factRsample)
}
>>>>>>> dbb4e3154a829c44cc223f4a4ca18fbc00e0ba13
