% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation-methods.R
\name{getAScons}
\alias{getAScons}
\title{Quantify feature conservation}
\usage{
getAScons(object, db = "phastCons", type = "exon", padding = 50)
}
\arguments{
\item{object}{factR object}

\item{db}{Database to extract sequence conservation. Can be "phastCons" (Default) or "phylop"}

\item{type}{Feature to quantify conservation. Can be one of the following:
\itemize{
 \item{"exon"}{ : Sequence conservation of the entire exon}
 \item{"flanks"}{ : Conservation of sequences flanking exons}
 \item{"upstream"}{ : Conservation of sequences upstream of exons}
 \item{"downstream"}{ : Conservation of sequences downstream of exons}
}}

\item{padding}{Additional width to pad the sequence by. For cons_type "exons" and "flanks", paddings will be added on both sides.}
}
\value{
factRObject with additional columns in ASE metadata.
}
\description{
Returns sequence conservation scores at the level of exons.
}
\examples{
data(factRsample)
factRsample <- getASCons(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
