% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listSupportedGenomes}
\alias{listSupportedGenomes}
\title{List supported genomes}
\usage{
listSupportedGenomes()
}
\value{
Dataframe of the IDs of genomes and its description
}
\description{
List supported genomes
}
\examples{
listSupportedGenomes()
}
