% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Plot}
\alias{Plot}
\title{Visualize RNA transcripts or translated proteins}
\usage{
Plot(object, ..., type = "transcripts", rescale_introns = FALSE, ncol = 1)
}
\arguments{
\item{object}{factRObject}

\item{...}{Optional: a list of features to plot. Input
can be a mixture of names from gene_id, gene_name or transcript_id metadata.
If missing, function will plot transcripts/proteins from the first 9 genes
of the custom transcriptome}

\item{type}{data to show. Can be "transcripts" or "proteins"}

\item{rescale_introns}{when plotting transcripts, whether to rescale introns}

\item{ncol}{number of columns to combine multiple feature plots to}
}
\value{
ggplot2 object. If multiple genes are detected, plots will be
combined using patchwork
}
\description{
Plots out transcript architecture or translated protein domains from
selected genes
}
\examples{
data(factRsample)
View(factRsample, "Selenop")
}
\author{
Fursham Hamid
}
