% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plotTranscripts}
\alias{plotTranscripts}
\alias{plotDomains}
\title{Visualize custom transcriptome and proteins}
\usage{
plotTranscripts(object, ..., rescale_introns = FALSE, ncol = 1)

plotDomains(object, ..., ncol = 1)
}
\arguments{
\item{object}{factRObject}

\item{...}{Optional: a list of features to plot. Input
can be a mixture of names from gene_id, gene_name or transcript_id metadata.
If missing, function will plot transcripts from the first 9 genes
of the custom transcriptome}

\item{rescale_introns}{when plotting transcripts, whether to rescale introns}

\item{ncol}{number of columns to combine multiple feature plots to}
}
\value{
ggplot2 object. If multiple genes are detected, plots will be
combined using patchwork
}
\description{
Plots out transcripts or protein domains from custom transcriptome
}
\examples{
# plot transcripts
data(factRsample)
plotTranscripts(factRsample, "Selenop")

# plot protein domains
factRsample <- runfactR(factRsample)
plotDomains(factRsample, "Prkaa1")
}
\author{
Fursham Hamid
}
