% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-methods.R
\name{predictDomains}
\alias{predictDomains}
\title{Predict protein domains}
\usage{
predictDomains(object, ..., database = "superfamily", ncores = 4)
}
\arguments{
\item{object}{factRObject}

\item{...}{Features to predict for. Can be gene_name, gene_id or transcript_id}

\item{database}{HMM database to query.}

\item{ncores}{Number of cores to run prediction on}
}
\value{
Updated factRObject
}
\description{
Predict protein domains
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- getAAsequence(factRsample)
factRsample <- predictDomains(factRsample, "Osmr")
}
