% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnmd-methods.R
\name{predictNMD}
\alias{predictNMD}
\title{Predict NMD-sensitive transcripts}
\usage{
predictNMD(object, NMD_threshold = 50, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{NMD_threshold}{Minimum distance between PTC and downstream exon-exon junction to trigger NMD (Default: 50)}

\item{verbose}{Whether to print out messages (Default: FALSE)}
}
\description{
Annotates NMD-sensitive transcripts based on Exon-Junction model.
By default, stop codons which are >50bp upstream of the last exon junction will
be annotated as NMD-sensitive
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- predictNMD(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
