% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{txRanges}
\alias{txRanges}
\alias{txData}
\title{Preview factR object}
\usage{
txRanges(object, ...)

txData(object)
}
\arguments{
\item{object}{factRObject}
}
\description{
A set of functions to view the contents of a factR object.
\itemize{
 \item{view(): }{view contents of custom transcriptome in spreadsheet-style format}
 \item{txRanges(): }{prints out transcript ranges}
 \item{txData(): }{prints out transcript metadata}
}
}
