% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{View}
\alias{View}
\alias{head}
\alias{tail}
\title{Preview custom transcriptome from factR object}
\usage{
View(object, ..., type = "transcripts", in_console = FALSE)

head(object, n = 6L)

tail(object, n = 6L)
}
\arguments{
\item{object}{factRObject}

\item{...}{Optional: a list of features to view. Input
can be a mixture of names from gene_id, gene_name or transcript_id metadata.
If missing, function will return all features}

\item{type}{data to show. Can be "transcripts" or "proteins"}

\item{in_console}{whether to print custom transcriptome in console}

\item{n}{an integer of the length to display}
}
\value{
GenomicRanges object
}
\description{
Test test
}
\author{
Fursham Hamid
}
