% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rangesData}
\alias{rangesData}
\alias{featData}
\title{Preview factR object}
\usage{
rangesData(object, ..., set = NULL)

featData(object, ..., set = NULL)
}
\arguments{
\item{object}{factRObject}
}
\description{
A set of functions to view the contents of a factR object.
\itemize{
 \item{view(): }{view contents of custom transcriptome in spreadsheet-style format}
 \item{txRanges(): }{prints out transcript ranges}
 \item{featureData(): }{prints out transcript metadata}
}
}
