% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{view}
\alias{view}
\alias{head}
\alias{tail}
\title{Preview factR object}
\usage{
view(object, ..., in_console = FALSE)

head(object, n = 6L)

tail(object, n = 6L)
}
\arguments{
\item{object}{factRObject}

\item{...}{Optional: a list of features to view. Input
can be a mixture of names from gene_id, gene_name or transcript_id metadata.
If missing, function will return all features}

\item{in_console}{whether to print custom transcriptome in console}

\item{n}{an integer of the length to display}

\item{type}{data to show. Can be "transcripts" or "proteins"}
}
\value{
GenomicRanges object
}
\description{
A set of functions to view the contents of a factR object.
\itemize{
 \item{view: }{view contents of custom transcriptome in spreadsheet-style format}
 \item{head: }{prints out the first n lines of custom transcriptome}
 \item{tail: }{prints out the last n lines of custom transcriptome}
}
}
\author{
Fursham Hamid
}
