% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factRObject-class.R
\name{runfactR}
\alias{runfactR}
\title{Run factR workflow}
\usage{
runfactR(object, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{verbose}{whether to display workflow progress}
}
\value{
Updated factRObject
}
\description{
This wrapper function will perform the core factR workflow which
include (1) building of CDSs, (2) predicting of NMD-sensitive transcripts
and (3) labelling of alternative splicing segments.
}
\examples{
data(factRsample)
factRsample <- runfactR(factRsample)
}
\seealso{
\code{\link{buildCDS}}, \code{\link{predictNMD}}, \code{\link{findAltSplicing}}
}
