% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{testASNMDevents}
\alias{testASNMDevents}
\title{Identify AS-NMD events}
\usage{
testASNMDevents(object)
}
\arguments{
\item{object}{factRObject}
}
\value{
Updated factRObject
}
\description{
This function will xxx
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- testASNMDevents(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
