% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnmd-methods.R
\name{testASNMDevents}
\alias{testASNMDevents}
\title{Identify AS-NMD events}
\usage{
testASNMDevents(object, verbose = FALSE)
}
\arguments{
\item{object}{factRObject}

\item{verbose}{Whether to print out messages (Default: FALSE)}
}
\value{
factRObject with updated ASE metadata
}
\description{
Pinpoints alternative splicing events underlying NMD sensitivity.
Metadata of AS events will be updated to include columns "ASNMDtype" and "ASNMD.in.cds".
Values of "ASNMDtype" can be Repressing (skipping of exon causes transcripts to be NMD-sensitive),
Stimulating (splicing of exon causes transcripts to be NMD-sensitive) or
NA (does not cause transcripts to NMD-sensitive).
}
\examples{
data(factRsample)
factRsample <- buildCDS(factRsample)
factRsample <- predictNMD(factRsample)
factRsample <- testASNMDevents(factRsample)
}
\seealso{
\code{\link{runfactR}}
}
